refactor: move constants out of config and into constants
