+++ test/integration/narrow/routes/farmer-apply/login.test.js
@@ -5,7 +5,7 @@ const expectLoginPage = require('../../../../utils/login-page-expect')
 const pageExpects = require('../../../../utils/page-expects')
 const expectPhaseBanner = require('../../../../utils/phase-banner-expect')
 const { notify: { templateIdFarmerApplyLogin }, serviceUri } = require('../../../../../app/config')
-const { farmerApply } = require('../../../../../app/config/user-types')
+const { farmerApply } = require('../../../../../app/constants/user-types')
 const uuidRegex = require('../../../../../app/config/uuid-regex')
 const loginTypes = require('../../../../../app/constants/login-types')
